requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
flask==3.0.3
flask-sqlalchemy==3.1.1
itsdangerous==2.2.0
jinja2==3.1.3
markupsafe==2.1.5
pycparser==2.22
pymysql==1.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
typing_extensions==4.11.0
werkzeug==3.0.2